# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/audit-logs": {
      "get": {
        "operationId": "Audit_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "actorId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "action",
            "required": false,
            "in": "query",
            "schema": {
              "example": "user.suspend",
              "type": "string"
            }
          },
          {
            "name": "targetType",
            "required": false,
            "in": "query",
            "schema": {
              "example": "user",
              "type": "string"
            }
          },
          {
            "name": "targetId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLogPageDto"
                }
              }
            }
          }
        },
        "summary": "Journal des actions des administrateurs, du plus récent au plus ancien",
        "tags": [
          "audit"
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "operationId": "Auth_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResultDto"
                }
              }
            }
          }
        },
        "summary": "Créer un compte apprenant et ouvrir une session",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/register/trainer": {
      "post": {
        "operationId": "Auth_registerTrainer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResultDto"
                }
              }
            }
          }
        },
        "summary": "Demander un compte formateur (validé ensuite par un admin)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "Auth_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResultDto"
                }
              }
            }
          }
        },
        "summary": "Se connecter avec email et mot de passe",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/two-factor/setup": {
      "post": {
        "operationId": "Auth_setUpTwoFactor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TwoFactorChallengeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TwoFactorSetupDto"
                }
              }
            }
          }
        },
        "summary": "Configurer la double authentification (première connexion d’un admin)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/two-factor/verify": {
      "post": {
        "operationId": "Auth_verifyTwoFactor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TwoFactorVerifyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResultDto"
                }
              }
            }
          }
        },
        "summary": "Valider le code de double authentification",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "Auth_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResultDto"
                }
              }
            }
          }
        },
        "summary": "Renouveler la session (le jeton est remplacé à chaque appel)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "Auth_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Fermer la session en cours",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "Auth_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserDto"
                }
              }
            }
          }
        },
        "summary": "Utilisateur connecté, son rôle et ses permissions",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/users": {
      "get": {
        "operationId": "Users_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Sans filtre, les comptes supprimés sont exclus.",
            "schema": {
              "$ref": "#/components/schemas/UserStatus"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "Clé du rôle.",
            "schema": {
              "example": "TRAINER",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Recherche dans le nom, le prénom et l’email.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPageDto"
                }
              }
            }
          }
        },
        "summary": "Lister les comptes (filtres : état, rôle, recherche)",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "Users_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "summary": "Consulter un compte",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "Users_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Supprimer un compte (données personnelles effacées)",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}/approve": {
      "post": {
        "operationId": "Users_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "summary": "Valider un compte formateur en attente",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}/suspend": {
      "post": {
        "operationId": "Users_suspend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "summary": "Suspendre un compte (ses sessions sont fermées)",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}/reactivate": {
      "post": {
        "operationId": "Users_reactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "summary": "Réactiver un compte suspendu",
        "tags": [
          "users"
        ]
      }
    },
    "/api/health": {
      "get": {
        "operationId": "Health_check",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthDto"
                }
              }
            }
          }
        },
        "summary": "État de l’API et de sa base de données",
        "tags": [
          "health"
        ]
      }
    }
  },
  "info": {
    "title": "API APGO",
    "description": "API unique du site, de l’administration et de l’application mobile APGO. Les erreurs suivent toutes le même format : statusCode, code, message, details, requestId.",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "apgo_access"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "AuditActorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "AuditLogDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "action": {
            "type": "string",
            "example": "user.suspend"
          },
          "targetType": {
            "type": "string",
            "example": "user"
          },
          "targetId": {
            "type": "string",
            "nullable": true
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          },
          "ip": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "actor": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/AuditActorDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "action",
          "targetType",
          "targetId",
          "metadata",
          "ip",
          "createdAt",
          "actor"
        ]
      },
      "AuditLogPageDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditLogDto"
            }
          }
        },
        "required": [
          "total",
          "page",
          "pageSize",
          "items"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "ama.kouassi@exemple.ci",
            "maxLength": 254
          },
          "password": {
            "type": "string",
            "minLength": 10,
            "maxLength": 128
          },
          "firstName": {
            "type": "string",
            "example": "Ama",
            "maxLength": 100
          },
          "lastName": {
            "type": "string",
            "example": "Kouassi",
            "maxLength": 100
          },
          "locale": {
            "type": "string",
            "enum": [
              "fr",
              "goa"
            ],
            "default": "fr"
          },
          "client": {
            "type": "string",
            "enum": [
              "web",
              "mobile"
            ],
            "default": "web"
          }
        },
        "required": [
          "email",
          "password",
          "firstName",
          "lastName"
        ]
      },
      "RoleSummaryDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "example": "LEARNER"
          },
          "name": {
            "type": "string",
            "example": "Apprenant"
          }
        },
        "required": [
          "key",
          "name"
        ]
      },
      "CurrentUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "locale": {
            "type": "string",
            "example": "fr"
          },
          "role": {
            "$ref": "#/components/schemas/RoleSummaryDto"
          },
          "permissions": {
            "type": "array",
            "description": "Pour adapter l’interface ; les droits restent vérifiés par l’API.",
            "items": {
              "type": "string",
              "enum": [
                "admin.access",
                "users.read",
                "users.manage",
                "admins.manage",
                "roles.manage",
                "audit.read",
                "settings.manage",
                "payments.switch",
                "payments.read",
                "media.manage",
                "courses.supervise",
                "courses.teach",
                "learning.access",
                "bible.manage",
                "support.handle",
                "certificates.reissue"
              ]
            }
          },
          "twoFactorEnabled": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "locale",
          "role",
          "permissions",
          "twoFactorEnabled"
        ]
      },
      "AuthResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "AUTHENTICATED",
              "TWO_FACTOR_REQUIRED",
              "TWO_FACTOR_SETUP_REQUIRED",
              "PENDING_APPROVAL"
            ]
          },
          "user": {
            "$ref": "#/components/schemas/CurrentUserDto"
          },
          "challengeToken": {
            "type": "string"
          },
          "accessToken": {
            "type": "string",
            "description": "Mobile uniquement."
          },
          "refreshToken": {
            "type": "string",
            "description": "Mobile uniquement."
          },
          "accessTokenExpiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "status"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "ama.kouassi@exemple.ci"
          },
          "password": {
            "type": "string"
          },
          "client": {
            "type": "string",
            "enum": [
              "web",
              "mobile"
            ],
            "default": "web"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "TwoFactorChallengeDto": {
        "type": "object",
        "properties": {
          "challengeToken": {
            "type": "string",
            "description": "Jeton reçu à la connexion (valable 5 minutes)."
          }
        },
        "required": [
          "challengeToken"
        ]
      },
      "TwoFactorSetupDto": {
        "type": "object",
        "properties": {
          "secret": {
            "type": "string",
            "description": "Secret à saisir à la main si le QR code ne peut pas être lu."
          },
          "otpauthUri": {
            "type": "string",
            "description": "Lien `otpauth://` à afficher en QR code."
          }
        },
        "required": [
          "secret",
          "otpauthUri"
        ]
      },
      "TwoFactorVerifyDto": {
        "type": "object",
        "properties": {
          "challengeToken": {
            "type": "string",
            "description": "Jeton reçu à la connexion (valable 5 minutes)."
          },
          "code": {
            "type": "string",
            "example": "123456",
            "description": "Code à 6 chiffres de l’application."
          },
          "client": {
            "type": "string",
            "enum": [
              "web",
              "mobile"
            ],
            "default": "web"
          }
        },
        "required": [
          "challengeToken",
          "code"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          },
          "client": {
            "type": "string",
            "enum": [
              "web",
              "mobile"
            ],
            "default": "web"
          }
        }
      },
      "UserStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "ACTIVE",
          "SUSPENDED",
          "DELETED"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserStatus"
              }
            ]
          },
          "locale": {
            "type": "string",
            "example": "fr"
          },
          "role": {
            "$ref": "#/components/schemas/RoleSummaryDto"
          },
          "twoFactorEnabled": {
            "type": "boolean"
          },
          "lastLoginAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "status",
          "locale",
          "role",
          "twoFactorEnabled",
          "lastLoginAt",
          "createdAt"
        ]
      },
      "UserPageDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserDto"
            }
          }
        },
        "required": [
          "total",
          "page",
          "pageSize",
          "items"
        ]
      },
      "HealthDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "ok"
          }
        },
        "required": [
          "status"
        ]
      }
    }
  }
}